requirements.txt
flask
numpy
pandas
scikit-learn
joblib
requests
python-dotenv
shap